refactored simple udt scripts
